Usability and performance improvements
Significantly speed up the process by copying Gallery2 thumbnails and
resizes wherever possible instead of regenerating them.  This requires
us to figure out the dimensions of the original G2 derivative and make
sure that it matches in some reasonable way.

To allow users to take advantage of this, calculate the optimal thumb
and resize size to set in G3 to match what was used in G2.  While
we're at it, give the user some idea of how much data is available in
G2 to import.