Fix up the titles in the default theme.  They've been broken for a while.  This fixes ticket #342.
The bug is that we were using $item instead of $theme->item().  But we
were also not special casing tags properly, and they are effectively
first class citizens (at least for now) so treat them properly.  Also,
set page_title by default in the theme so that we don't have to do an
empty() check on it (makes the theme easier to read) and move the
title out of Tags_Controller so that the theme has more control over
it.

Signed-off-by:  <unostar@danalan.info>