2nd attempt at gracefully dealing with sort columns that contain nulls.  This time around, do a query to determine whether or not the sort column has nulls in it.  If it doesn't, then use our comparators as before.
There are NULLs in the sort column, so we can't use MySQL comparators.
Fall back to iterating over every child row to get to the current one.
This can be wildly inefficient for really large albums, but it should
be a rare case that the user is sorting an album with null values in
the sort column.

Fixes #627