Revert "Use call_user_func instead of call_user_func_array as we don't need to pass the parameters by reference." This reverts commit 75cc4962a2c945e3e3ad37fda3b5de32d512b929.
Revert "Make sure that calls to the theme::get_var() have sane defaults in the event that the theme does not supply them."
This reverts commit 51f2b547672ab69a7a04d37b723c6afbeb584a59.

Revert "Update the installer to reflect the new location of the theme related options."
This reverts commit 678e22996d73b96833c48d94751344a6bd8f2899.

Revert "Create theme::get_var(), theme::set_var() methods to set the options of the active site theme.  Change all refrences to theme options to use these methods.  Update the version number of Gallery to 20 and move any them related options to the be stored under the active theme."
This reverts commit 26114972c3388f065220b94a0d5962f20a6ccd0c.

Revert "Currently Admin_Theme_Options controller assumes that all the themes will provide the same values. This change corrects that assumption and moves the management of the theme options, including creating the form and updating the theme options into the theme."
This reverts commit 1692ee130887a6ad1ba68aa34a96ad36161600f9.