Move model business logic out of album::create() and into Item_Model::save().  This makes creating albums similar to editing them and makes it difficult to create an album poorly.  I expect to be able to remove a lot of code from the photo and movie helper because it's duplicated here.
In order to do this, I refactored ORM_MPTT::add_to_parent() into
ORM_MPTT::save() so we now add it to the parent when we do save.  This
allows us to call save() only once which saves a database call per
add.

The Albums_Controller logic is roughly the same as before.

Haven't updated the tests yet, they're going to fail miserably since
many of them depend on album::create() which is now gone.