Add a "weight" column to the module table.  This allows us to specify module ordering, which is currently being done in the moduleorder contrib module.
By default, the weight will be the same as the id of the row which
means that new modules will get added at the end of the list.  This is
covered in the upgrade case as well.

The one gotcha is that we need to make sure that we don't try to sort
by the weight column if the gallery module version is < 32, which is
something we haven't done before.