First step in setting up version checking.
We now have two types of packages (release, git).  Instead of using
constants, we now have gallery::version_string() which returns the
current version string.  If you're on a release package, then the
version string looks like:

  3.0 (Santa Fe)

If you're on a git package, then the version string looks like this:

  3.0.1 (branch 3.0.x build 3)

We track the build number in a new file in the gallery3 root called
BUILD_NUMBER which we will update periodically with the latest
build number for each branch.