Overhaul the Admin > Content > Comments page.  Use a tabbed interface for consistency with other parts of the admin UI (ie: G2 import).
The UI is now Ajax driven because that's how jQuery tabs works best in
this case.  This requires a little finagling in JS to get the
paginator links to load properly.  While I'm in there, add a paginator
at the top of the table, make the buttons in the delete tab sane and
smooth scroll back to the top of the tab when switching pages.

Fixes #1702