Encapsulate the context data into the Display_Context. The data can be set individually by using key value pairs; or, the data values can be set using an assocative array.
The context name is passed as an argument into the factory method.  If the context name is null, we assume we are retreiving the context (on display) and get it from the session. If context name has been supplied, then we are initializing the context and create context based on the name. (i.e. a name of 'item' will instantiate a 'Item_Display_Context' class.

The method 'display_context' has been added which returns the page values that depend on which context the item is being displayed.  The Display_Context class also contains a method to build the build crumb list.