Support development environment configuration via Vagrant.
This commit allows using Vagrant[1] to configure the development
environment. To start hacking on Gallery, all you have to do is:

 $ git clone git@github.com:gallery/gallery3.git
 $ cd gallery3
 $ sudo echo "192.168.33.30\tgallery3.local" >> /etc/hosts
 $ vagrant up

 (open http://gallery3.local in a web browser)

The Vagrantfile will instruct Vagrant to download a preconfigured
VM which contains all development dependencies and tools, assign it
to a host-only network IP address and boot it. The directory which
contains the source code will be mounted using NFS at /gallery3.

You can easily access the VM to, for example, run unit tests:

 $ vagrant ssh
 $ cd /gallery3
 $ php index.php test

[1] http://www.vagrantup.com