Reverting the ORM::$_sorting change because it doesn't do quite what we want when it comes to dealing with descendants.
Basically, if you do:

    $album = Test::random_album_unsaved();
    $album->sort_column = "view_count";
    $album->sort_order = "DESC";
    $album->save();
    $album->descendants->find_all();

(captured in Item_Model_Test::test_descendants_default_to_albums_sort_order)

what happens is that ORM_MPTT::get() returns a new instance of
Model_Item with the left and right pointer constrained:

      return ORM::factory($this->_model_name)
        ->where("left_ptr", ">", $this->left_ptr)
        ->where("right_ptr", "<", $this->right_ptr);

but this is a default instance that doesn't have any of the other
values from the original instance ($album in the example above) so it
doesn't the ORM::$_sorting member variable correctly, so
ORM::_loaded_result() won't pick it up.

I didn't see an obvious way to resolve this so I wrote a test and
rolled this back for now.  Feel free to roll it forward again with an
implementation that makes the test pass.  :-)

--

Revert "Ensure that the Model_Item default sorting is reset after save(), too."
This reverts commit 6a4233b5c9c9ae1565eaf597747cf01ba29bed92.

Revert "Use ORM::$_sorting to set default sorting, remove $item->unsorted_children."
This reverts commit ffee4501219e0887b9a4a5bd09c9ca6308a7a597.